ppu: Implement STAT and LYC interrupts
Modified step() to request LCD_STAT interrupt on mode transitions to 0/1/2 (first step before cycle increment, on horizontal line increment, and ly = 154) when the corresponding enable bit is set.

Added check for STAT interrupts to track which STAT modes have already triggered or latched an interrupt this frame. This is to avoid repeated requests to LCD_STAT while the condition is still true