joypad: Add P1 register methods, update addressBus & gbEmulator
Game Boy reads controller input via P1 register, which returns an 8 bit number in the format 11 XY ZZZZ (little-endian order).

Bits 7-6 are always set to 1.
Bit 5 selects from buttons (A/B/Select/Start) if set to 0.
Bit 4 selects from directions (Right/Left/Up/Down) if set to 0.
Bits 3-0 are set to 0 for "pressed" and 1 for "not pressed".

Inside the joypad class, directionalState and buttonState are used to map an active-high state from press and release listener events (1 if pressed and 0 if not pressed)

pressButton(btn) sets the corresponding bit internally to 1 if it was 0. If state changed from 0 to 1, request InterruptType.JOYPAD

releaseButton(btn) clears the corresponding bit to 0 (not pressed internally)

On CPU reads from P1, selection bits XY are used to compute the lower nibble ZZZZ based on the selected state. readP1LowerNibble checks which direction is selected, clears bits for pressed buttons and returns the lower nibble in an active-low state.

On CPU writes to P1, only selection bits 5 and 4 are stored into the IO register.