ppu: implement per-scanline BG index buffer
Now bgIndexLine is used to store the 2bit BG/window color index for each pixel on the current scanline. Window pixels overwrite both the framebuffer color and index buffer values.

Also changed the sprite OBJ-to-BG priority to now be index-based. If priorityBehindBg is clear, always draw over BG. If it's set, only draw where bgIndexLine[x] is 0.

This follows the OAM priority attribute: If the priority bit is set to 0, the sprite will be drawn behind the background and window layers. While if it's 1 the sprite will be drawn on top except in cases where the BG index is the transparent color 0

This should fix edge cases in which sprites were incorrectly rendered in front of solid BG tiles or vice versa.