app: save RAM snapshots from battery powered cartridges to localStorage
Upon closing/refreshing the tab or resetting the game, saveSRAMToLocalStorage in App.tsx calls getSRAMSnapshot, makes a copy of the cartridge's RAM and saves it to localStorage encoded in base64.

On ROM loads it looks for an unique key (made in cartridge.getSaveKey()) with the game's internal title, cartridge type and checksum. If there is one, it decodes it back to a string. Then cartridge.loadSRAMSnapshot() restores the content back to cartridge RAM before starting the emulator.

localStorage has a 5MB limit per domain and each save is about 10 to 50KB so this should be alright for now.