ppu: fix tearing issues by using a front buffer to return completed frames
useGameBoyEmulator hook holds the RAF loop to drive the emulation's logic. While GameboyScreen handles the rendering by copying pixels from ppu.getFramebuffer() when frameReady is set at currentScanlineLY === 144

The issue is that the framebuffer returned was being written concurrently from the emulator's logic, and it gets reused for the next frame. If emulation is fast enough and advances past VBlank, the top scanlines of the buffer can be overwritten by the next frame before the UI gets to copy the previous one to the screen.

Fixed by copying the completed frame into a dedicated frontbuffer on Vblank and returning that from getFramebuffer() so the UI always reads a stable snapshot